fix(fsext): panic on fastwalk (#1122)
this seems to fix a panic I've been getting some times.

bubbletea eats part of the trace, so I'm not sure it's a race, but
judging by the code, and the fact that fastwalk creates multiple
goroutines, it makes sense.

i've been using this for the past few hours and haven't encountered the
error since.

Signed-off-by: Carlos Alexandro Becker <caarlos0@users.noreply.github.com>